Move post-authorization endpoint to separate controller (fixes redirect loop)
