+++ src/NuGet.Status/Views/Authorized/Index.cshtml
@@ -14,7 +14,7 @@
     <div class="row">
         <div class="col-xs-12">
             @{
-                var updated = TempData[AdminController.UpdatedTempDataKey];
+                var updated = TempData[AuthorizedController.UpdatedTempDataKey];
             }
             @if (updated != null)
             {
