+++ src/NuGet.Status/Views/Authorized/_EditStatusEventMessage.cshtml
